+++ src/Core/Type/TypeComparator.php
@@ -22,6 +22,7 @@ use Gskema\TypeSniff\Core\Type\Common\UnionType;
 use Gskema\TypeSniff\Core\Type\Declaration\NullableType;
 use Gskema\TypeSniff\Core\Type\DocBlock\ClassStringType;
 use Gskema\TypeSniff\Core\Type\DocBlock\DoubleType;
+use Gskema\TypeSniff\Core\Type\DocBlock\KeyValueType;
 use Gskema\TypeSniff\Core\Type\DocBlock\ThisType;
 use Gskema\TypeSniff\Core\Type\DocBlock\TrueType;
 use Gskema\TypeSniff\Core\Type\DocBlock\TypedArrayType;
@@ -158,6 +159,8 @@ class TypeComparator
 
         $flatDocTypes = $docType instanceof UnionType ? $docType->getTypes() : [$docType];
         foreach ($flatDocTypes as $flatDocType) {
+            $flatDocType = $flatDocType instanceof KeyValueType ? $flatDocType->getType() : $flatDocType;
+
             $flatDocTypeClass = get_class($flatDocType);
             $coveredFnTypeClasses = static::$coveredFnTypeClassMap[$flatDocTypeClass] ?? [];
             $coveredFnTypeClasses[] = $flatDocTypeClass;